python/viperEAFIT/requirements.txt
astroid==2.2.5
Django==2.1.7
django-bulma==0.5.6
django-multiselectfield==0.1.8
isort==4.3.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==2.3.1
pylint-django==2.0.6
pylint-plugin-utils==0.5
pytz==2018.9
six==1.12.0
typed-ast==1.3.1
wrapt==1.11.1
